fix bugs: (1) not removing self-loops when directed option selected, (2) change to use int value instead of float fixes kron21 graph coloring. bugs left: (1) spmspvSimple does not interact well with mxv int due to atomic lacking an int option, (2) spmspvMerge fails for directed and undirected graph coloring

+++ example/ggc.cu
@@ -23,7 +23,7 @@ int main( int argc, char** argv )
 {
   std::vector<graphblas::Index> row_indices;
   std::vector<graphblas::Index> col_indices;
-  std::vector<float> values;
+  std::vector<int> values;
   graphblas::Index nrows, ncols, nvals;
 
   // Parse arguments
@@ -65,7 +65,7 @@ int main( int argc, char** argv )
     CHECK( desc.toggle(graphblas::GrB_INP1) );
 
   // Matrix A
-  graphblas::Matrix<float> a(nrows, ncols);
+  graphblas::Matrix<int> a(nrows, ncols);
   values.clear();
   values.resize(nvals, 1.f);
   CHECK( a.build(&row_indices, &col_indices, &values, nvals, GrB_NULL,
